WIP RTF parsing.
This is the complicated result of parsing a semi-structured and very idiosyncratic RTF document containing all the ICD-9 definitions from top-level down to billable codes. This replaces the web scraping from icd9data.com, which remains a good resource for checking the results. This is a step towards fixing correct range generation in some corner cases, because we now include all the sub-headings which are not billable, and didn't show up before.